added rtk in understandingroute to handle routes instead of state
