+++ Routing/understandingroute/src/App.js
@@ -1,7 +1,6 @@
-import { Navigate, Route, Routes, useRoutes } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import "./App.css";
 import About from "./components/About";
-import Home from "./components/Home";
 import Layout from "./components/Layout";
 import NotFound from "./components/NotFound";
 import ProductDetails from "./components/ProductDetails";
